% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph.R
\name{bargraph}
\alias{bargraph}
\title{Make a bar graph}
\usage{
bargraph(data, x_var, y_var)
}
\arguments{
\item{data}{This represents the dataset}

\item{x_var}{This is the column that goes on the x axis}

\item{y_var}{This is the column that goes on the y axis}
}
\value{
bar This should now be a bargraph
}
\description{
Using two columns from a dataset to make a bar graph
}
