% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_volume.R
\name{calc_shrub_vol}
\alias{calc_shrub_vol}
\title{Calculate volume using H\emph{L}W}
\usage{
calc_shrub_vol(height, length, width)
}
\arguments{
\item{height}{height of object}

\item{length}{length of object}

\item{width}{width of object}
}
\value{
density of the object
}
\description{
Using Height, Length and Width to calculate volume by multiplying the three number
}
