% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_function.R
\name{fun_function}
\alias{fun_function}
\title{Accidental function. YOLO}
\usage{
fun_function(data)
}
\arguments{
\item{data}{data you would like to ruin}
}
\value{
data that is unusable
}
\description{
Data is made unusable because data is all numeric with no columns, legit random numbers
}
