% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_data.R
\name{function_graph}
\alias{function_graph}
\title{Graphs X vs Y on a scatter plot}
\usage{
function_graph(data, column, column_1)
}
\arguments{
\item{data}{data set being used}

\item{column}{column for x axis}

\item{column_1}{column for y axis}
}
\value{
graph of x vs y
}
\description{
Creates a scatter plot used to show trends between x and y in a data set
}
