% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Make a linear model}
\usage{
linear_model(df, dependent, independent_cols)
}
\arguments{
\item{df}{This is the dataset}

\item{dependent}{This is the numeric dependent column of your choice from the dataset}

\item{independent_cols}{These are the independent predictor columns of your choice from the dataset}
}
\value{
return_summary This should give you the linear model
}
\description{
Use a numeric column and two predictor columns to make a linear model
}
