% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_function.R
\name{lm_function}
\alias{lm_function}
\title{Linear model for crabs}
\usage{
lm_function(df, dependent, independent)
}
\arguments{
\item{df}{data set being used}

\item{dependent}{x in the lm}

\item{independent}{y in the lm}
}
\value{
Linear model chart
}
\description{
compares X~Y comlumns using an linear model
}
