% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply_2}
\alias{multiply_2}
\title{Multiply column}
\usage{
multiply_2(data, column, new_column)
}
\arguments{
\item{data}{This is the dataset}

\item{column}{This is the column of your choice from your dataset}

\item{new_column}{This is the new column generated}
}
\value{
new There should now be a new column that is multiplied by 2
}
\description{
Multiply a column in a dataset by 2 and then generate a new column
}
