% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_data.R
\name{mutate_data}
\alias{mutate_data}
\title{mutates the data by reorganizing}
\usage{
mutate_data(data, column_names)
}
\arguments{
\item{data}{data set being used}

\item{column}{column used for reordering}
}
\value{
data reordered
}
\description{
with alter data ordering by a chosen column
}
