% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{plot}
\alias{plot}
\title{Make a scatter plot}
\usage{
plot(data, column_one, column_two)
}
\arguments{
\item{data}{This represents the dataset}

\item{column_one}{Column of your choice from a dataset}

\item{column_two}{Second column of your choice from a dataset}
}
\value{
scatter This should be a scatter plot
}
\description{
Use two columns of your choice from a dataset to then make a scatter plot
}
