% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_NAs.R
\name{remove_nas}
\alias{remove_nas}
\title{Clean data}
\usage{
remove_nas(data)
}
\arguments{
\item{data}{This represents the dataset}
}
\value{
clean This dataset shouldn't have NAs now
}
\description{
Use all of the columns in a dataset to clean them by removing NAs
}
