% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{removes in na's from  data set}
\usage{
clean_data(data)
}
\arguments{
\item{data}{data to be cleaned}
}
\value{
data with no na's
}
\description{
well remove an missing value/ NA rows to aloow for functions to be completed
}
