% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_function.R
\name{lm_function}
\alias{lm_function}
\title{Calculate volume using H\emph{L}W}
\usage{
lm_function(df, column, column_2)
}
\arguments{
\item{df}{data set being used}

\item{column}{x in the lm}

\item{column_2}{y in the lm}
}
\value{
Linear model chart
}
\description{
Using Height, Length and Width to calculate volume by multiplying the three number
}
