% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_data.R
\name{mutate_data}
\alias{mutate_data}
\title{mutates the data and creates a new column}
\usage{
mutate_data(data, column, new_column)
}
\arguments{
\item{data}{data set being used}

\item{column}{column used for mutating}

\item{new_column}{name of mutated column}
}
\value{
Linear model chart
}
\description{
with alter data and create a new column with an assigned name
}
